+++ src/snowflake/cli/_plugins/nativeapp/sf_sql_facade.py
@@ -40,6 +40,7 @@ from snowflake.cli._plugins.nativeapp.sf_facade_exceptions import (
     UserScriptError,
     handle_unclassified_error,
 )
+from snowflake.cli._plugins.stage.manager import StageManager
 from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.constants import ObjectType
 from snowflake.cli.api.errno import (
@@ -251,7 +252,7 @@ class SnowflakeSQLFacade:
     def create_version_in_package(
         self,
         package_name: str,
-        stage_fqn: str,
+        path_to_version_directory: str,
         version: str,
         label: str | None = None,
         role: str | None = None,
@@ -259,7 +260,7 @@ class SnowflakeSQLFacade:
         """
         Creates a new version in an existing application package.
         @param package_name: Name of the application package to alter.
-        @param stage_fqn: Stage fully qualified name.
+        @param path_to_version_directory: Path to artifacts on the stage to create a version from.
         @param version: Version name to create.
         @param [Optional] role: Switch to this role while executing create version.
         @param [Optional] label: Label for this version, visible to consumers.
@@ -274,6 +275,9 @@ class SnowflakeSQLFacade:
         with_label_clause = (
             f"label={to_string_literal(label)}" if label is not None else ""
         )
+        using_clause = (
+            f"using {StageManager.quote_stage_name(path_to_version_directory)}"
+        )
 
         action = "register" if available_release_channels else "add"
 
@@ -282,7 +286,7 @@ class SnowflakeSQLFacade:
                 f"""\
                     alter application package {package_name}
                         {action} version {version}
-                        using @{stage_fqn}
+                        {using_clause}
                         {with_label_clause}
                 """
             )
@@ -326,7 +330,7 @@ class SnowflakeSQLFacade:
     def add_patch_to_package_version(
         self,
         package_name: str,
-        stage_fqn: str,
+        path_to_version_directory: str,
         version: str,
         patch: int | None = None,
         label: str | None = None,
@@ -335,7 +339,7 @@ class SnowflakeSQLFacade:
         """
         Add a new patch, optionally a custom one, to an existing version in an application package.
         @param package_name: Name of the application package to alter.
-        @param stage_fqn: Stage fully qualified name.
+        @param path_to_version_directory: Path to artifacts on the stage to create a version from.
         @param version: Version name to create.
         @param [Optional] patch: Patch number to create.
         @param [Optional] label: Label for this patch, visible to consumers.
@@ -352,11 +356,12 @@ class SnowflakeSQLFacade:
             f"\nlabel={to_string_literal(label)}" if label is not None else ""
         )
         patch_query = f"{patch}" if patch else ""
+        using_clause = StageManager.quote_stage_name(path_to_version_directory)
         add_patch_query = dedent(
             f"""\
                  alter application package {package_name}
                      add patch {patch_query} for version {version}
-                     using @{stage_fqn}{with_label_clause}
+                     using {using_clause}{with_label_clause}
              """
         )
         with self._use_role_optional(role):
@@ -632,7 +637,7 @@ class SnowflakeSQLFacade:
         self,
         name: str,
         install_method: SameAccountInstallMethod,
-        stage_fqn: str,
+        path_to_version_directory: str,
         role: str,
         warehouse: str,
         debug_mode: bool | None,
@@ -643,7 +648,7 @@ class SnowflakeSQLFacade:
 
         @param name: Name of the application object
         @param install_method: Method of installing the application
-        @param stage_fqn: FQN of the stage housing the application artifacts
+        @param path_to_version_directory: Path to directory in stage housing the application artifacts
         @param role: Role to use when creating the application and provider-side objects
         @param warehouse: Warehouse which is required to create an application object
         @param debug_mode: Whether to enable debug mode; None means not explicitly enabled or disabled
@@ -658,7 +663,7 @@ class SnowflakeSQLFacade:
 
         with self._use_role_optional(role), self._use_warehouse_optional(warehouse):
             try:
-                using_clause = install_method.using_clause(stage_fqn)
+                using_clause = install_method.using_clause(path_to_version_directory)
                 upgrade_cursor = self._sql_executor.execute_query(
                     f"alter application {name} upgrade {using_clause}",
                 )
@@ -728,7 +733,7 @@ class SnowflakeSQLFacade:
         name: str,
         package_name: str,
         install_method: SameAccountInstallMethod,
-        stage_fqn: str,
+        path_to_version_directory: str,
         role: str,
         warehouse: str,
         debug_mode: bool | None,
@@ -741,7 +746,7 @@ class SnowflakeSQLFacade:
         @param name: Name of the application object
         @param package_name: Name of the application package to install the application from
         @param install_method: Method of installing the application
-        @param stage_fqn: FQN of the stage housing the application artifacts
+        @param path_to_version_directory: Path to directory in stage housing the application artifacts
         @param role: Role to use when creating the application and provider-side objects
         @param warehouse: Warehouse which is required to create an application object
         @param debug_mode: Whether to enable debug mode; None means not explicitly enabled or disabled
@@ -763,7 +768,7 @@ class SnowflakeSQLFacade:
             )
             authorize_telemetry_clause = f" AUTHORIZE_TELEMETRY_EVENT_SHARING = {str(should_authorize_event_sharing).upper()}"
 
-        using_clause = install_method.using_clause(stage_fqn)
+        using_clause = install_method.using_clause(path_to_version_directory)
         with self._use_role_optional(role), self._use_warehouse_optional(warehouse):
             try:
                 create_cursor = self._sql_executor.execute_query(