+++ RELEASE-NOTES.md
@@ -25,6 +25,7 @@
   * `snow app release-directive unset`
 * Add support for release channels feature in native app version creation/drop.
 * `snow app version create` now returns version, patch, and label in JSON format.
+* Add an optional `stage_subdirectory` field to `application package` entity. When specified, application artifacts are uploaded to this subdirectory instead of the root of the application package's stage.
 
 ## Fixes and improvements
 * Fixed crashes with older x86_64 Intel CPUs.